Proper handling for security context continuation
In the case that the security context is not established after one loop. one would need to bind the context against the TCP connection, continue and wait for the client response. Unfortunately, Tomcat does not offer an API to access the context and store objects in in. Ultimately, we support stateless authentication only.